refactor(data): import concrete modules instead of directory barrels
Importing a single type from a directory barrel pulls in that directory's
whole module graph. In packages/data this closed 16 import cycles, the
dominant one being domain <-> update:

  domain/pin/pin-updater.ts  -> '../../update'   (barrel)
  update/index.ts            -> re-exports data-updater
  update/data-updater.ts     -> '../domain'      (barrel)
  domain/index.ts            -> re-exports pin-updater

pin-updater only wants the Updater interface; the barrel is what closes
the loop. Repointing these imports at concrete modules takes the package
from 16 cycles to 0, so no-circular moves from 'warn' to 'error'.

The eight domain/*/index.ts barrels are deleted: each had exactly one
importer, domain/index.ts, which now re-exports concrete files. No
consumer deep-imports those paths -- every import of this package targets
its root. The root barrel stays; that is the package's public interface.

Public API is unchanged: 32 exported symbols before and after, compared
via the TypeScript checker against dist/index.d.ts.
