+++ packages/data/src/domain/asset/asset-updater.ts
@@ -1,5 +1,5 @@
 import { Storage } from '../../storage';
-import { Updater } from '../../update';
+import { Updater } from '../../update/updater';
 import { Asset } from './asset';
 
 export class AssetUpdater implements Updater {
