--- packages/data/src/domain/asset/index.ts
@@ -1,3 +1,0 @@
-export * from './asset-updater';
-export * from './asset-service';
-export * from './asset';
