+++ packages/data/src/domain/index.ts
@@ -1,8 +1,28 @@
-export * from './text';
-export * from './language';
-export * from './asset';
-export * from './station';
-export * from './pin';
-export * from './tour';
-export * from './server';
-export * from './share';
+export * from './asset/asset';
+export * from './asset/asset-service';
+export * from './asset/asset-updater';
+
+export * from './language/language';
+export * from './language/language-service';
+export * from './language/language-updater';
+
+export * from './pin/pin-service';
+export * from './pin/pin-updater';
+
+export * from './server/server-service';
+export * from './server/server-updater';
+
+export * from './share/share-service';
+export * from './share/share-updater';
+
+export * from './station/station';
+export * from './station/station-service';
+export * from './station/station-updater';
+
+export * from './text/text';
+export * from './text/text-service';
+export * from './text/text-updater';
+
+export * from './tour/tour';
+export * from './tour/tour-service';
+export * from './tour/tour-updater';
