--- packages/data/src/domain/language/index.ts
@@ -1,3 +1,0 @@
-export * from './language-updater';
-export * from './language-service';
-export * from './language';
