--- packages/data/src/domain/pin/index.ts
@@ -1,2 +1,0 @@
-export * from './pin-updater';
-export * from './pin-service';
