--- packages/data/src/domain/server/index.ts
@@ -1,2 +1,0 @@
-export * from './server-service';
-export * from './server-updater';
