--- packages/data/src/domain/share/index.ts
@@ -1,2 +1,0 @@
-export * from './share-service';
-export * from './share-updater';
