--- packages/data/src/domain/station/index.ts
@@ -1,3 +1,0 @@
-export * from './station';
-export * from './station-updater';
-export * from './station-service';
