--- packages/data/src/domain/text/index.ts
@@ -1,3 +1,0 @@
-export * from './text';
-export * from './text-updater';
-export * from './text-service';
