--- packages/data/src/domain/tour/index.ts
@@ -1,3 +1,0 @@
-export * from './tour';
-export * from './tour-updater';
-export * from './tour-service';
